Create README.md
I thrive on experimenting with different concepts and have been commended for inspiring and developing unique and thought-provoking ideas. I have cultivated Leadership, Creativity and Digital Marketing skills through a range of practical experience. These strengths, combined with my passion for the industry have created a strong foundation to meet the needs of the Marketing Officer opportunity. I strive to contribute to an energetic and team-driven work culture.